// projects/mmb-web-app/src/app/shared/types/interest.model.ts
export class Interest {
    user: string;
    email: string;
    username: string;
    uuid_of_notification: string;
    career: string;
    career_slug: string;
    event: string;
    event_slug: string;
    initaitve: string;
    initiative_slug: string;
    opportunity: string;
    opportunity_slug: '';
    source: string;
    source_slug: string;
    event_type: string;
}